feat: updatepost in repository and changed post answers count

+++ app/src/main/java/com/muratdayan/lessonline/domain/model/firebasemodels/Post.kt
@@ -12,7 +12,7 @@ data class Post(
     val photoUri: String="",
     val comment:String="",
     var likeCount:Int =0,
-    val answerCount:Int=0,
+    val answerCount:Long=0,
     var likedByUsers: List<String> = emptyList(),
     val postAnswers: List<String> = emptyList(),
     var isBookmarked : Boolean = false,
